+++ server/index.js
@@ -1,5 +1,11 @@
 import app from './app.js'
+import { connectToDb } from './db.js'
+import dotenv from 'dotenv'
+dotenv.config()
 
 const port = process.env.PORT || 4002
 
-app.listen(port, () => console.log(`App running at http://localhost:${port}/`))
+app.listen(port, () => {
+    console.log(`App running at http://localhost:${port}/`)
+    connectToDb()
+})
