+++ server/routes/games_routes.js
@@ -1,5 +1,6 @@
 import express from 'express'
-import { GameModel } from '../db.js'
+import { RatingModel, ReviewModel, GameModel } from '../models/UserModel.js'
+
 
 const router = express.Router()
